pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]

[project]
name = "cflac-carbon-aware-training"
version = "0.1.0"

# -------------------------------------------------------------------------
# Exhaustive dependency list – all imported modules MUST appear here.
# -------------------------------------------------------------------------

dependencies = [
  "torch>=2.0",
  "timm>=0.9.12",
  "hydra-core>=1.3",
  "wandb>=0.16",
  "optuna>=3.4",
  "datasets>=2.15",
  "huggingface_hub>=0.18",
  "pandas>=2.0",
  "matplotlib>=3.8",
  "seaborn>=0.13",
  "pynvml>=11.5",
  "scikit-learn>=1.3",
  "numpy>=1.26",
  "requests>=2.31",
  "scipy>=1.11"
]
